Added User Model and Authentication

+++ backend/controllers/goalController.js
@@ -1,12 +1,13 @@
 const asyncHandler = require("express-async-handler")
 const Goal = require("../models/goalmodel")
 const { default: mongoose } = require("mongoose")
+const User = require("../models/userModel")
 
 // @desc    get goals
 // @route   GET api/goals
 // @access  Private
 const getGoals = asyncHandler(async(req, res) => {
-    const goals = await Goal.find({})
+    const goals = await Goal.find({ user: req.user.id})
     res.status(200).json(goals)
 })
 
@@ -17,7 +18,10 @@ const setGoal = asyncHandler(async (req, res) => {
     if(!req.body.text) {
         res.status(400).json({message: "Please add a text field"})
     }
-    const goal = await Goal.create({text: req.body.text})
+    const goal = await Goal.create({
+        text: req.body.text,
+        user: req.user.id
+    })
     res.status(200).json(goal)
 })
 
@@ -25,14 +29,33 @@ const setGoal = asyncHandler(async (req, res) => {
 // @route   PUT api/goals/:id
 // @access  Private
 const updateGoal = asyncHandler(async (req, res) => {
-    const goal = await Goal.findById(req.params.id) 
-    if(!goal) {
-        res.status(400)
-        throw new Error("Goal not found")
+  const goal = await Goal.findById(req.params.id)
+
+  if (!goal) {
+    res.status(400)
+    throw new Error('Goal not found')
+  }
+
+ const user = await User.findById(req.user.id)
+    // Check for user
+    if(!user) {
+        res.status(401)
+        throw new Error("User not found")
     }
-    const updatedGoal = await Goal.findByIdAndUpdate(req.params.id, req.body, {new: true})
-    res.status(200).json(updatedGoal)
+
+  // Make sure the logged in user matches the goal user
+  if (goal.user.toString() !== user.id) {
+    res.status(401)
+    throw new Error('User not authorized')
+  }
+
+  const updatedGoal = await Goal.findByIdAndUpdate(req.params.id, req.body, {
+    new: true,
+  })
+
+  res.status(200).json(updatedGoal)
 })
+
 
 // @desc    delete goal
 // @route   DELETE api/goals
@@ -42,6 +65,21 @@ const deleteGoal = asyncHandler(async (req, res) => {
     if(!goal) {
         res.status(400).json("Goal not found")
     }
+
+    const user = await User.findById(req.user.id)
+    // Check for user
+    if(!user) {
+        res.status(401)
+        throw new Error("User not found")
+    }
+
+    // Make sure that the logged user matches the goal user
+    if(goal.user.toString() !== user.id) {
+        res.status(401)
+        throw new Error("User not authorized")
+    }
+
+    // Delete the Goal
     await goal.deleteOne()
 
     res.status(200).json({id: req.params.id})
